sx126x: Fix typing of debug print
These would pass ostime_t (aka uint32_t) typed values to a printf that
expects `%d` (long int) parameters. This is often the same, but there
are exist some platforms where `long int` is 64-bit.

Similarly, an osticks2us result was passed to `%F`, which expects
`u4_t`, so make that explicit too.

+++ lmic/radio-sx127x.c
@@ -861,7 +861,7 @@ static void rxlorasingle (void) {
     // warn about delayed rx
     if( rxtime - now < 0 ) {
         debug_printf("WARNING: rxtime is %ld ticks in the past! (ramp-up time %ld ms / %ld ticks)\r\n",
-                     now - rxtime, osticks2ms(now - t0), now - t0);
+                     (long)(now - rxtime), (long)osticks2ms(now - t0), (long)(now - t0));
     }
 }
 
@@ -994,7 +994,7 @@ static void rxfsk (bool rxcontinuous) {
         ostime_t now = os_getTime();
         if (LMIC.rxtime - now < 0) {
             debug_printf("WARNING: rxtime is %ld ticks in the past! (ramp-up time %ld ms / %ld ticks)\r\n",
-                         now - LMIC.rxtime, osticks2ms(now - t0), now - t0);
+                         (long)(now - LMIC.rxtime), (long)osticks2ms(now - t0), (long)(now - t0));
         }
         hal_waitUntil(LMIC.rxtime);
     }
